Add backend integration testing infrastructure
