+++ tests/setup.js
@@ -1,7 +1,6 @@
-require('../models/User');
+require('dotenv').config();
+require('../models');
 
 const mongoose = require('mongoose');
-const keys = require('../config/keys');
 
 mongoose.Promise = global.Promise;
-mongoose.connect(keys.mongoURI, { useNewUrlParser: true });
